+++ backend/de.metas.adempiere.adempiere/base/src/main/java/de/metas/security/IUserRolePermissionsDAO.java
@@ -161,4 +161,6 @@ public interface IUserRolePermissionsDAO extends ISingletonService
 	 * @return true if given user has a role where he/she is an administrator, according to {@link IUserRolePermissions#isSystemAdministrator()}
 	 */
 	boolean isAdministrator(ClientId clientId, UserId adUserId, LocalDate date);
+
+	void deleteUserOrgAccessByUserId(final UserId userId);
 }
