+++ backend/de.metas.adempiere.adempiere/base/src/main/java/org/adempiere/service/IValuePreferenceDAO.java
@@ -2,6 +2,7 @@ package org.adempiere.service;
 
 import java.util.Properties;
 
+import de.metas.user.UserId;
 import org.compiere.model.I_AD_Preference;
 
 import de.metas.util.ISingletonService;
@@ -51,4 +52,9 @@ public interface IValuePreferenceDAO extends ISingletonService
 	 */
 	void save(Properties ctx, String attribute, Object value, int adClientId, int adOrgId, int adUserId, int adWindowId);
 
+	/**
+	 *
+	 * @param userId
+	 */
+	void deleteUserPreferenceByUserId(final UserId userId);
 }
