+++ backend/de.metas.adempiere.adempiere/base/src/main/java/org/adempiere/service/impl/ValuePreferenceDAO.java
@@ -2,6 +2,7 @@ package org.adempiere.service.impl;
 
 import java.util.Properties;
 
+import de.metas.user.UserId;
 import org.adempiere.ad.dao.IQueryBL;
 import org.adempiere.model.InterfaceWrapperHelper;
 import org.adempiere.model.PlainContextAware;
@@ -40,6 +41,7 @@ import de.metas.util.Services;
 public class ValuePreferenceDAO implements IValuePreferenceDAO
 {
 	private final transient Logger logger = LogManager.getLogger(getClass());
+	private final IQueryBL queryBL = Services.get(IQueryBL.class);
 
 	@Override
 	public boolean remove(final Properties ctx, final String attribute, final int adClientId, final int adOrgId, final int adUserId, final int adWindowId)
@@ -130,5 +132,14 @@ public class ValuePreferenceDAO implements IValuePreferenceDAO
 		{
 			return value.toString();
 		}
+	}
+
+	@Override
+	public void deleteUserPreferenceByUserId(final UserId userId)
+	{
+		queryBL.createQueryBuilder(I_AD_Preference.class)
+				.addEqualsFilter(I_AD_Preference.COLUMNNAME_AD_User_ID, userId)
+				.create()
+				.delete();
 	}
 }
